@@ -1,37 +1,38 @@
 {
-  "name": "zavrsni_rad",
-  "version": "0.1.0",
+  "name": "robinzon_rezervacija",
+  "version": "1.0.0",
   "private": true,
   "homepage": "http://www.robinzonlucica.eu/",
   "dependencies": {
-    "@auth0/auth0-react": "^1.4.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "date-fns": "^2.21.1",
-    "gh-pages": "^3.1.0",
-    "leaflet": "^1.7.1",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.5.2",
-    "react-cookie-consent": "^6.2.3",
-    "react-datepicker": "^3.7.0",
-    "react-dom": "^17.0.2",
+    "@auth0/auth0-react": "^1.10.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.2.0",
+    "axios": "^0.27.2",
+    "date-fns": "^2.28.0",
+    "gh-pages": "^4.0.0",
+    "leaflet": "^1.8.0",
+    "moment": "^2.29.3",
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.4.0",
+    "react-cookie-consent": "^7.4.1",
+    "react-datepicker": "^4.8.0",
+    "react-dom": "^18.1.0",
     "react-helmet": "^6.1.0",
-    "react-icons": "^4.2.0",
-    "react-image-gallery": "^1.0.9",
-    "react-leaflet": "^3.1.0",
-    "react-loader-spinner": "^4.0.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-icons": "^4.3.1",
+    "react-image-gallery": "^1.2.8",
+    "react-leaflet": "^4.0.0",
+    "react-loader-spinner": "^5.1.4",
+    "react-phone-number-input": "^3.2.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "react-scroll-up": "^1.3.7",
-    "react-timecalendar": "^2.0.1",
-    "react-to-print": "^2.12.4",
-    "react-toastify": "^7.0.4",
-    "styled-components": "^5.2.3",
+    "react-timecalendar": "^2.1.3",
+    "react-to-print": "^2.14.7",
+    "react-toastify": "^9.0.1",
+    "styled-components": "^5.3.5",
     "validator": "^13.7.0",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -52,5 +53,10 @@
     "not dead",
     "not op_mini all"
   ],
-  "proxy": "http://localhost:3001"
+  "proxy": "http://localhost:3001",
+  "devDependencies": {
+    "autoprefixer": "^10.4.7",
+    "postcss": "^8.4.14",
+    "tailwindcss": "^3.0.24"
+  }
 }
