@@ -1,12 +1,11 @@
 import React from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { Auth0Provider } from "@auth0/auth0-react";
-require("dotenv").config();
 const Auth0ProviderWithHistory = ({ children }) => {
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const onRedirectCallback = (appState) => {
-    history.push(appState?.returnTo || window.location.pathname);
+    navigate(appState?.returnTo || window.location.pathname);
   };
 
   return (
