@@ -1,19 +1,14 @@
 import React from "react";
 import { BrowserRouter as Router } from "react-router-dom";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import Auth0ProviderWithHistory from "./auth/auth0";
-
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
   <Router>
     <Auth0ProviderWithHistory>
       <App />
     </Auth0ProviderWithHistory>
-  </Router>,
-  document.getElementById("root")
+  </Router>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
