@@ -1,15 +1,12 @@
 import React from "react";
-import { Route } from "react-router-dom";
 import { withAuthenticationRequired } from "@auth0/auth0-react";
 import { Loading } from "../components/index";
 
-const ProtectedRoute = ({ component, ...args }) => (
-  <Route
-    component={withAuthenticationRequired(component, {
-      onRedirecting: () => <Loading />,
-    })}
-    {...args}
-  />
-);
+const ProtectedRoute = ({ component, ...args }) => {
+  const Cp = withAuthenticationRequired(component, {
+    onRedirecting: () => <Loading />,
+  });
+  return <Cp {...args} />;
+};
 
 export default ProtectedRoute;
