@@ -2,5 +2,13 @@ import Pocetna from "./pocetna";
 import Footer from "./footer";
 import Rezervacija from "./rezervacija";
 import RezervacijaAdmin from "./rezadmin";
-import ErrorPage from "./errorpage"
-export { Pocetna, Footer, Rezervacija, RezervacijaAdmin, ErrorPage };
+import ErrorPage from "./errorpage";
+import Statistics from "./statistics";
+export {
+  Pocetna,
+  Footer,
+  Rezervacija,
+  RezervacijaAdmin,
+  ErrorPage,
+  Statistics,
+};
